moved io-api to examples
