+++ mini_al/_examples/io_api/capture.go
@@ -1,4 +1,4 @@
-package malgo
+package io_api
 
 import (
 	"context"
